code complete

+++ src/main/java/com/edureka/application/rest/model/Movie.java
@@ -2,18 +2,18 @@ package com.edureka.application.rest.model;
 
 import java.io.Serializable;
 
-public class Movie implements Serializable{
+public class Movie implements Serializable {
 
 	/**
 	 * 
 	 */
 	private static final long serialVersionUID = -2731430661826302300L;
-	
+
 	private String title;
-	
+
 	private String alsoViewed;
-	
-	private Long count;
+
+	private Integer count;
 
 	public String getTitle() {
 		return title;
@@ -31,12 +31,12 @@ public class Movie implements Serializable{
 		this.alsoViewed = alsoViewed;
 	}
 
-	public Long getCount() {
+	public Integer getCount() {
 		return count;
 	}
 
-	public void setCount(Long count) {
+	public void setCount(Integer count) {
 		this.count = count;
 	}
-	
+
 }
